requirements.txt
# LangChain ecosystem
langchain==1.2.15
langchain-core==1.3.0
langchain-community==0.4.1
langchain-groq==1.1.2
langchain-chroma==1.1.0
langchain-huggingface==1.2.2
langchain-text-splitters==1.1.2

# Vector store
chromadb==1.5.8

# Embeddings & models
sentence-transformers==5.4.1
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.11.0+cpu
transformers==5.5.4
tokenizers==0.22.2

# PDF loading
pypdf==6.10.2

# LLM provider
groq==0.37.1

# API layer
fastapi==0.136.0
uvicorn==0.44.0
python-multipart==0.0.26

# UI
gradio==6.12.0

# Evaluation
ragas==0.4.3
datasets==4.8.4

# Utilities
python-dotenv==1.2.2
pydantic==2.13.2